# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Зашли почитать документацию? Жаль",
        "title": "Quest Service API",
        "contact": {}
    },
    "host": "localhost:5000",
    "paths": {
...
        "/api/v1/quest-service/quest/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение всех опросов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QuestResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/all/with-statuses": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение всех опросов со статусами",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QuestWithStatuses"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/by-user/with-statuses/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опросов по ID пользователя со статусами",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QuestWithStatusesForUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/by-user/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опросов по ID пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QuestWithIndicators"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Создание опроса",
                "parameters": [
                    {
                        "description": "Body для создания опроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QuestCreate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token (с биркой)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Quest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/hide/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Скрытие опроса по ID (используйте как удаление)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Quest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/with-indicators/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опроса c его вопросами по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestWithIndicators"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/with-users-and-indicators/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опроса c его вопросами и пользователями по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestWithUsersAndIndicators"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/with-users/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опроса c его пользователями по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestWithUsers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/quest-service/quest/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Получение опроса по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.IndicatorWithCategory": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available": {
                    "type": "boolean"
                },
                "categoryName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fromRole": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idСategory": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "toRole": {
                    "type": "string"
                }
            }
        },
        "models.Quest": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "endAt": {
                    "type": "integer"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "startAt": {
                    "type": "integer"
                }
            }
        },
        "models.QuestCreate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endAt": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "startAt": {
                    "type": "integer"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "models.QuestResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.QuestWithIndicators": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IndicatorWithCategory"
                    }
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.QuestWithStatuses": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TeamWithStatus"
                    }
                }
            }
        },
        "models.QuestWithStatusesForUser": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "isPass": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TeamWithStatus"
                    }
                }
            }
        },
        "models.QuestWithUsers": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.QuestWithUsersAndIndicators": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "endAt": {
                    "type": "string"
                },
                "idQuest": {
                    "type": "integer"
                },
                "idQuestTemplate": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IndicatorWithCategory"
                    }
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.TeamWithStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserWithStatus"
                    }
                }
            }
        },
...
        "models.UserWithStatus": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "boolean"
                }
            }
        },
...
    }
}